refactor(FON-468): apply PR review feedback on feature-first structure
