--- apps/client/src/constants/mimetypes.constants.ts
@@ -1,12 +1,0 @@
-export const ACCEPTED_MIME_TYPES = ['application/pdf', 'image/jpeg', 'image/png'];
-
-export const EMBEDDED_SCREENSHOTS_ACCEPTED_MIME_TYPES = [
-  'image/jpeg',
-  'image/png',
-  'image/gif',
-  'image/webp',
-];
-
-export const TRANSPARENCES_ACCEPTED_MIME_TYPES = [
-  'application/vnd.openxmlformats-officedocument.spreadsheetml.sheet',
-];
