--- apps/client/src/constants/sg-anchor-attributes.ts
@@ -1,5 +1,0 @@
-export const SECRETARIAT_GENERAL_ANCHOR_ATTRIBUTES = {
-  dashboard: '/secretariat-general',
-  nouvelleTransparence: '/secretariat-general/nouvelle-transparence',
-  transparence: (id: string) => `/secretariat-general/transparence/session/${id}`,
-} as const;
