+++ apps/client/src/features/nomination-files-table/components/cells/magistrat-details/MagistratDetails.tsx
@@ -5,7 +5,7 @@ import { FormattedMessage } from 'react-intl';
 import { labels } from '@/constants/labels.constants';
 import { useIsSgNavigation } from '@/features/auth/hooks/roles.hook';
 import { ObservationCard } from '@/features/observations/components/ObservationCard';
-import { reportHtmlIds } from '@/features/reports/dom/html-ids';
+import { reportHtmlIds } from '@/features/reports/constants/html-ids.constants';
 import { formatBiography, formatObservers } from '@/features/reports/utils/formatters';
 import { FormattedBirthDate } from '@/i18n/components';
 import { useIntlPositionDuration } from '@/i18n/hooks';
