+++ apps/client/src/features/reports/components/Biography.tsx
@@ -1,5 +1,5 @@
 import { labels } from '@/constants/labels.constants';
-import { reportHtmlIds } from '@/features/reports/dom/html-ids';
+import { reportHtmlIds } from '@/features/reports/constants/html-ids.constants';
 
 import { Card } from './Card';
