+++ apps/client/src/features/reports/components/Summary.tsx
@@ -1,11 +1,11 @@
 import { SideMenu, type SideMenuProps } from '@codegouvfr/react-dsfr/SideMenu';
 import { type FC, useCallback, useEffect, useMemo, useState } from 'react';
 
+import { reportHtmlIds } from '@/features/reports/constants/html-ids.constants';
 import { SUMMARY_SECTIONS } from '@/features/reports/constants/summary-sections.constants';
-import { reportHtmlIds } from '@/features/reports/dom/html-ids';
-import { scrollToSummarySection } from '@/features/reports/dom/scroll-to-summary-section';
-import { summaryScrollListenersFactory } from '@/features/reports/dom/summary-scroll-listeners';
 import { useObservedSections } from '@/features/reports/hooks/useObservedSections';
+import { scrollToSummarySection } from '@/features/reports/utils/scroll-to-summary-section';
+import { summaryScrollListenersFactory } from '@/features/reports/utils/summary-scroll-listeners';
 
 export type SummaryProps = {
   observers: string[] | null;
