+++ apps/client/src/features/reports/utils/scroll-to-summary-section.ts
