+++ apps/client/src/features/reports/utils/summary-section-observer-factory.ts
