+++ apps/client/src/features/sessions/components/ManageSession.tsx
@@ -4,7 +4,7 @@ import { format } from 'date-fns';
 import { defineMessage } from 'react-intl';
 import { generatePath, Link, useLocation } from 'react-router';
 
-import type { BreadcrumbVM } from '@/models/breadcrumb-vm.model';
+import type { BreadcrumbVM } from '@/shared/ui/Breadcrumb';
 import { Breadcrumb } from '@/shared/ui/Breadcrumb';
 import { DataTable, useDataTable, useQueryDataTableState } from '@/shared/ui/data-table';
 import { dateOnlyToDate } from '@/utils/date-only.util';
