--- apps/client/src/models/breadcrumb-vm.model.ts
@@ -1,9 +1,0 @@
-import type { To } from 'react-router';
-
-export type BreadcrumbVM = {
-  currentPageLabel: string;
-  segments: {
-    label: string;
-    to: To;
-  }[];
-};
